+++ GarbageCan/Data/Models/Boosters/Booster.cs
@@ -1,4 +1,4 @@
-namespace GarbageCan.Data.Models
+namespace GarbageCan.Data.Models.Boosters
 {
     public class Booster
     {
