+++ GarbageCan/Data/Models/Boosters/UserBooster.cs
@@ -1,4 +1,4 @@
-﻿namespace GarbageCan.Data.Models
+﻿namespace GarbageCan.Data.Models.Boosters
 {
     public class UserBooster
     {
