+++ GarbageCan/Data/Models/XP/User.cs
@@ -1,4 +1,4 @@
-﻿namespace GarbageCan.Data.Models
+﻿namespace GarbageCan.Data.Models.XP
 {
     public class User
     {
